requirements.txt
flask==3.0.2
sentence-transformers==2.7.0
faiss-cpu==1.7.4
transformers==4.38.2
torch==2.2.1
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1
tqdm==4.66.2